Moved RepositoryFilterUtils methods to RepositoryFilter
