+++ eva-server/src/main/java/uk/ac/ebi/eva/server/ws/RegionWSServer.java
@@ -40,12 +40,10 @@ import uk.ac.ebi.eva.lib.repository.VariantEntityRepository;
 import uk.ac.ebi.eva.lib.utils.DBAdaptorConnector;
 import uk.ac.ebi.eva.lib.utils.MultiMongoDbFactory;
 import uk.ac.ebi.eva.lib.utils.RepositoryFilter;
-import uk.ac.ebi.eva.lib.utils.RepositoryFilterUtils;
 import uk.ac.ebi.eva.server.Utils;
 
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -85,7 +83,7 @@ public class RegionWSServer extends EvaWSServer {
         MultiMongoDbFactory.setDatabaseNameForCurrentThread(DBAdaptorConnector.getDBName(species));
 
         List<RepositoryFilter> filters =
-                RepositoryFilterUtils.getRepositoryFilters(maf, polyphenScore, siftScore, studies, consequenceType);
+                RepositoryFilter.getRepositoryFilters(maf, polyphenScore, siftScore, studies, consequenceType);
 
         List<Region> regions = Region.parseRegions(regionId);
