+++ eva-server/src/main/java/uk/ac/ebi/eva/server/ws/VariantWSServer.java
@@ -39,7 +39,6 @@ import uk.ac.ebi.eva.lib.repository.VariantEntityRepository;
 import uk.ac.ebi.eva.lib.utils.DBAdaptorConnector;
 import uk.ac.ebi.eva.lib.utils.MultiMongoDbFactory;
 import uk.ac.ebi.eva.lib.utils.RepositoryFilter;
-import uk.ac.ebi.eva.lib.utils.RepositoryFilterUtils;
 import uk.ac.ebi.eva.server.Utils;
 
 import javax.servlet.http.HttpServletResponse;
@@ -96,7 +95,7 @@ public class VariantWSServer extends EvaWSServer {
                                                            alternate);
         } else {
             List<RepositoryFilter> filters =
-                    RepositoryFilterUtils.getRepositoryFilters(maf, polyphenScore, siftScore, studies, consequenceType);
+                    RepositoryFilter.getRepositoryFilters(maf, polyphenScore, siftScore, studies, consequenceType);
 
             List<String> excludeMapped = exclude.stream().map(e -> Utils.getApiToMongoDocNameMap().get(e)).collect(
                     Collectors.toList());
